update local to repo

+++ src/uet/oop/bomberman/entities/Entity.java
@@ -9,10 +9,10 @@ import uet.oop.bomberman.graphics.Sprite;
 
 public abstract class Entity {
     protected double x;
-    protected int y;
+    protected double y;
     protected Image img;
 
-    public Entity( double x, int y, Image img) {
+    public Entity( double x, double y, Image img) {
         this.x = x;
         this.y = y;
         this.img = img;
